MINOR: Port RestoreIntegrationTest to the new protocol (#20347)
The test uses regular consumer to commit offsets. New protocol requires
a streams consumer since we are using streams groups, otherwise we run
into group ID conflicts.

Followed the addition of the KafkaAdmin interface for setting offsets, a
Kafka Admin client is created and used the interface in to set the
committed offsets instead of instantiating a consumer.

Also enable all tests for stream new protocol.

Reviewers: Alieh Saeedi<asaeedi@confluent.io>, Kirk True
 <ktrue@confluent.io>, Matthias Sax <mjsax@apache.org>, Bill Bejeck
 <bbejeck@apache.org>